parser/requirements.txt
requests==2.28.1
pydantic==1.10.2 
psycopg2-binary==2.9.5 
python-dotenv==0.21.0